step 2: move satin_column, extract validation_models
